refactor main, inject logic to DBProcessor
